+++ prestashop/ntresellerclub/sql/uninstall.sql
@@ -3,6 +3,7 @@ DROP TABLE IF EXISTS `PREFIX_ntresellerclub_log`;
 DROP TABLE IF EXISTS `PREFIX_ntresellerclub_notice`;
 DROP TABLE IF EXISTS `PREFIX_ntresellerclub_billing_event`;
 DROP TABLE IF EXISTS `PREFIX_ntresellerclub_price`;
+DROP TABLE IF EXISTS `PREFIX_ntresellerclub_ssl_product_mapping`;
 DROP TABLE IF EXISTS `PREFIX_ntresellerclub_hosting_product_mapping`;
 DROP TABLE IF EXISTS `PREFIX_ntresellerclub_cart_domain`;
 DROP TABLE IF EXISTS `PREFIX_ntresellerclub_service`;